Fix segfault over Ubuntu (maybe for other distro too)
the function in origin script only has two args, while it calling for three args.
I removed the undeclared arg `boot` , added a condition to display the splash if called when booting.
I don't know if it is expected behavior, at least after this commit it works on my machine without giving sigsegv to journalctl.